Some refactorings made in several files
