+++ app/[lang]/arbswap/(auth)/login/page.tsx
@@ -1,22 +1,22 @@
 'use client'
 
-import { loginSchema, TLoginSchema } from '@/types/authValidation.types'
+import { loginSchema, TLoginSchema } from '@/types/auth'
 
 import React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
-import { loginDictionary } from '@/localesContent'
+import { loginDictionary } from '@/constants/localesContent'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Cookies from 'js-cookie'
 import { useForm } from 'react-hook-form'
 
 import Button from '@/components/Common/Button'
-import { useUserStore } from '@/hooks/useStore'
+import { userStore } from '@/hooks/useStore'
 import { loginUser } from '@/services/ApiService'
 
 const Login = ({ params }: { params: { lang: string } }) => {
-  const { setIsAuthenticated } = useUserStore()
+  const { setIsAuth } = userStore()
   const router = useRouter()
   const {
     register,
@@ -28,15 +28,15 @@ const Login = ({ params }: { params: { lang: string } }) => {
     try {
       // ApiService function pass data, return response
       const response = await loginUser(data)
-      const responseData = response?.data
 
       if (response && (response.status < 200 || response.status >= 300)) {
         // response status is not 2xx
         alert('Login failed!')
+        console.log('Problem')
       } else {
-        const token = responseData.access_token
+        const token = response.access_token
         Cookies.set('token', token)
-        setIsAuthenticated()
+        setIsAuth()
         router.push('/arbswap/dashboard')
       }
     } catch (e) {
