+++ components/CustomNavbar/ArbSwap/index.tsx
@@ -4,9 +4,9 @@ import React, { useState } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import { arbswapNavbarLinks } from '@/constants'
-import { arbswapNavbarDictionary } from '@/localesContent'
+import { arbswapNavbarDictionary } from '@/constants/localesContent'
 
-import { useUserStore } from '@/hooks/useStore'
+import { userStore } from '@/hooks/useStore'
 
 interface ArbSwapNavbarProps {
   readonly locale: string
@@ -14,7 +14,7 @@ interface ArbSwapNavbarProps {
 
 function ArbSwapNavbar({ locale }: ArbSwapNavbarProps) {
   const [nav, setNav] = useState(false)
-  const { isAuthenticated } = useUserStore()
+  const { isAuth } = userStore()
 
   const commonStyles =
     'block h-1 rounded-sm bg-white transition-all duration-300 ease-out -translate-y-0.5'
@@ -65,7 +65,7 @@ function ArbSwapNavbar({ locale }: ArbSwapNavbarProps) {
           <Link className="px-2 lg:px-5" href="/arbswap/support">
             {arbswapNavbarDictionary[locale]?.support}
           </Link>
-          {isAuthenticated ? (
+          {isAuth ? (
             <Link
               className="rounded-full bg-gradient-to-r from-purple-600 via-blue-500 to-green-600 px-3 py-2 lg:px-5"
               href="/arbswap/dashboard"
